+++ docs/CHANGELOG.md
@@ -8,6 +8,7 @@ Released on: ???
 - Moved Water Hotspots & Bayou overlay into 2 separate overlays, because Water Hotspots might be popular outside of Bayou. **Please re-enable in settings and adjust position if you used it!**
 - Renamed Legion & Bobbin' Time tracker and its settings to Nearby entities counter, to fit new purpose of showing various entity counters. **Please re-enable in settings and adjust position if you used it!**
 - Added Chumcap buckets to Nearby entities counter [disabled by default].
+- Track Agatha's Contest rewards in the Fishing profit tracker. This includes Forest Essence and Agatha's coupons based on rank earned during contest.
 - Added "Items" settings section with the following functionalities [all disabled by default]:
   - Highlight slots with trash enchanted books. Book names are configurable. You can use it to quickly find books to throw away or insta sell.
   - Highlight Kat GUI slot with red color if you try to give her Epic Megalodon. _For those who sometimes puts pets into Kat instead of George by mistake c:_